refactored reducer placement
